Move migration datatypes into `matrix-sdk-libolm-migration`
